requirements.txt
Django==2.0.2
django-cors-headers==2.1.0
django-mysql==2.2.0
djangorestframework==3.7.7
mysqlclient==1.3.12
pkg-resources==0.0.0
pytz==2017.3
